requirements.txt
# Core Data Science
pandas>=2.0.0
numpy~=1.26.0
scikit-learn>=1.3.0
scikit-surprise>=1.1.3 # For Matrix Factorization/SVD

# High-Performance Search
faiss-cpu>=1.7.4 # Use faiss-gpu if you have a compatible GPU

# MLOps & Serving
mlflow>=2.7.1
fastapi>=0.100.0
uvicorn[standard]
pydantic>=2.0.0

# Development and Utilities
pytest>=7.0.0
tqdm>=4.66.1